refactor(IT-Cotato#185): 컨트롤러, 서비스 분리
