+++ src/main/java/com/cotato/kampus/domain/board/api/BoardController.java
@@ -1,20 +1,16 @@
 package com.cotato.kampus.domain.board.api;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.cotato.kampus.domain.board.application.BoardService;
 import com.cotato.kampus.domain.board.api.response.BoardListResponse;
 import com.cotato.kampus.domain.board.api.response.BoardWithDescriptionResponse;
-import com.cotato.kampus.domain.board.api.response.FavoriteBoardResponse;
 import com.cotato.kampus.domain.board.api.response.HomeBoardAndPostPreviewResponse;
 import com.cotato.kampus.domain.board.api.response.UniversityBoardResponse;
-import com.cotato.kampus.domain.post.dto.response.BoardCategoryFindResponse;
 import com.cotato.kampus.global.common.dto.DataResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -22,7 +18,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 
-@Tag(name = "게시판(BoardEntity) API", description = "게시판 관련 API(게시글 API는 Post)")
+@Tag(name = "게시판(Board) API", description = "게시판 관련 API")
 @RestController
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
 @RequestMapping("/v1/api/boards")
@@ -81,44 +77,5 @@ public class BoardController {
 				boardService.getBoard(boardId)
 			)
 		));
-	}
-
-	@PostMapping("/favorite/{boardId}")
-	@Operation(summary = "게시판 즐겨찾기 추가", description = "특정 게시판을 즐겨찾기에 추가합니다.")
-	public ResponseEntity<DataResponse<FavoriteBoardResponse>> addFavoriteBoard(
-		@PathVariable Long boardId
-	) {
-		return ResponseEntity.ok(DataResponse.from(
-			FavoriteBoardResponse.of(
-				boardService.addFavoriteBoard(boardId)
-			)
-		));
-	}
-
-	@DeleteMapping("/favorite/{boardId}")
-	@Operation(summary = "게시판 즐겨찾기 삭제", description = "특정 게시판을 즐겨찾기에서 삭제합니다.")
-	public ResponseEntity<DataResponse<FavoriteBoardResponse>> removeFavoriteBoard(
-		@PathVariable Long boardId
-	) {
-		return ResponseEntity.ok(DataResponse.from(
-			FavoriteBoardResponse.of(
-				boardService.removeFavoriteBoard(boardId)
-			)
-		));
-	}
-
-	@GetMapping("/boards/{boardId}/categories")
-	@Operation(summary = "게시판에 적용되는 카테고리 조회",
-		description = "boardId에 해당하는 게시판에 적용되는 카테고리를 조회합니다.")
-	public ResponseEntity<DataResponse<BoardCategoryFindResponse>> findCategories(
-		@PathVariable Long boardId
-	) {
-		return ResponseEntity.ok(
-			DataResponse.from(
-				BoardCategoryFindResponse.from(
-					boardService.findCategories(boardId)
-				)
-			)
-		);
 	}
 }
